refactor(channels): extract runtime state helpers
