+++ src/channels/mod.rs
@@ -35,6 +35,7 @@ pub mod nostr;
 pub mod notion;
 pub mod qq;
 pub mod reddit;
+pub(crate) mod runtime_state;
 pub mod session_backend;
 pub mod session_sqlite;
 pub mod session_store;
@@ -112,9 +113,18 @@ use std::fmt::Write;
 use std::path::{Path, PathBuf};
 use std::process::Command;
 use std::sync::atomic::AtomicBool;
-use std::sync::{Arc, Mutex, OnceLock};
-use std::time::{Duration, Instant, SystemTime};
+use std::sync::{Arc, Mutex};
+use std::time::{Duration, Instant};
 use tokio_util::sync::CancellationToken;
+
+use self::runtime_state::{
+    clear_sender_history, config_file_stamp, conversation_history_key, conversation_memory_key,
+    followup_thread_id, get_route_selection, interruption_scope_key, mark_sender_for_new_session,
+    maybe_apply_runtime_config_update, parse_runtime_command, resolve_provider_alias,
+    resolved_default_model, resolved_default_provider, runtime_config_store,
+    runtime_defaults_from_config, runtime_defaults_snapshot, set_route_selection,
+    take_pending_new_session, ChannelRuntimeDefaults, RuntimeConfigState,
+};
 
 /// Observer wrapper that forwards tool-call events to a channel sender
 /// for real-time threaded notifications.
@@ -270,33 +280,6 @@ struct ModelCacheEntry {
     models: Vec<String>,
 }
 
-#[derive(Debug, Clone)]
-struct ChannelRuntimeDefaults {
-    default_provider: String,
-    model: String,
-    temperature: f64,
-    api_key: Option<String>,
-    api_url: Option<String>,
-    reliability: crate::config::ReliabilityConfig,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-struct ConfigFileStamp {
-    modified: SystemTime,
-    len: u64,
-}
-
-#[derive(Debug, Clone)]
-struct RuntimeConfigState {
-    defaults: ChannelRuntimeDefaults,
-    last_applied_stamp: Option<ConfigFileStamp>,
-}
-
-fn runtime_config_store() -> &'static Mutex<HashMap<PathBuf, RuntimeConfigState>> {
-    static STORE: OnceLock<Mutex<HashMap<PathBuf, RuntimeConfigState>>> = OnceLock::new();
-    STORE.get_or_init(|| Mutex::new(HashMap::new()))
-}
-
 const SYSTEMD_STATUS_ARGS: [&str; 3] = ["--user", "is-active", "R.A.I.N..service"];
 const SYSTEMD_RESTART_ARGS: [&str; 3] = ["--user", "restart", "R.A.I.N..service"];
 const OPENRC_STATUS_ARGS: [&str; 2] = ["R.A.I.N.", "status"];
@@ -407,40 +390,6 @@ impl InFlightTaskCompletion {
             return;
         }
         self.notify.notified().await;
-    }
-}
-
-fn conversation_memory_key(msg: &traits::ChannelMessage) -> String {
-    // Include thread_ts for per-topic memory isolation in forum groups
-    match &msg.thread_ts {
-        Some(tid) => format!("{}_{}_{}_{}", msg.channel, tid, msg.sender, msg.id),
-        None => format!("{}_{}_{}", msg.channel, msg.sender, msg.id),
-    }
-}
-
-fn conversation_history_key(msg: &traits::ChannelMessage) -> String {
-    // Include reply_target for per-channel isolation (e.g. distinct Discord/Slack
-    // channels) and thread_ts for per-topic isolation in forum groups.
-    match &msg.thread_ts {
-        Some(tid) => format!(
-            "{}_{}_{}_{}",
-            msg.channel, msg.reply_target, tid, msg.sender
-        ),
-        None => format!("{}_{}_{}", msg.channel, msg.reply_target, msg.sender),
-    }
-}
-
-fn followup_thread_id(msg: &traits::ChannelMessage) -> Option<String> {
-    msg.thread_ts.clone().or_else(|| Some(msg.id.clone()))
-}
-
-fn interruption_scope_key(msg: &traits::ChannelMessage) -> String {
-    match &msg.interruption_scope_id {
-        Some(scope) => format!(
-            "{}_{}_{}_{}",
-            msg.channel, msg.reply_target, msg.sender, scope
-        ),
-        None => format!("{}_{}_{}", msg.channel, msg.reply_target, msg.sender),
     }
 }
 
@@ -706,319 +655,6 @@ fn strip_tool_result_content(text: &str) -> String {
     }
 
     cleaned.to_string()
-}
-
-fn supports_runtime_model_switch(channel_name: &str) -> bool {
-    matches!(channel_name, "telegram" | "discord" | "matrix")
-}
-
-fn parse_runtime_command(channel_name: &str, content: &str) -> Option<ChannelRuntimeCommand> {
-    if !supports_runtime_model_switch(channel_name) {
-        return None;
-    }
-
-    let trimmed = content.trim();
-    if !trimmed.starts_with('/') {
-        return None;
-    }
-
-    let mut parts = trimmed.split_whitespace();
-    let command_token = parts.next()?;
-    let base_command = command_token
-        .split('@')
-        .next()
-        .unwrap_or(command_token)
-        .to_ascii_lowercase();
-
-    match base_command.as_str() {
-        "/models" => {
-            if let Some(provider) = parts.next() {
-                Some(ChannelRuntimeCommand::SetProvider(
-                    provider.trim().to_string(),
-                ))
-            } else {
-                Some(ChannelRuntimeCommand::ShowProviders)
-            }
-        }
-        "/model" => {
-            let model = parts.collect::<Vec<_>>().join(" ").trim().to_string();
-            if model.is_empty() {
-                Some(ChannelRuntimeCommand::ShowModel)
-            } else {
-                Some(ChannelRuntimeCommand::SetModel(model))
-            }
-        }
-        "/new" => Some(ChannelRuntimeCommand::NewSession),
-        _ => None,
-    }
-}
-
-fn resolve_provider_alias(name: &str) -> Option<String> {
-    let candidate = name.trim();
-    if candidate.is_empty() {
-        return None;
-    }
-
-    let providers_list = providers::list_providers();
-    for provider in providers_list {
-        if provider.name.eq_ignore_ascii_case(candidate)
-            || provider
-                .aliases
-                .iter()
-                .any(|alias| alias.eq_ignore_ascii_case(candidate))
-        {
-            return Some(provider.name.to_string());
-        }
-    }
-
-    None
-}
-
-fn resolved_default_provider(config: &Config) -> String {
-    config
-        .default_provider
-        .clone()
-        .unwrap_or_else(|| "openrouter".to_string())
-}
-
-fn resolved_default_model(config: &Config) -> String {
-    config
-        .default_model
-        .clone()
-        .unwrap_or_else(|| "anthropic/claude-sonnet-4.6".to_string())
-}
-
-fn runtime_defaults_from_config(config: &Config) -> ChannelRuntimeDefaults {
-    ChannelRuntimeDefaults {
-        default_provider: resolved_default_provider(config),
-        model: resolved_default_model(config),
-        temperature: config.default_temperature,
-        api_key: config.api_key.clone(),
-        api_url: config.api_url.clone(),
-        reliability: config.reliability.clone(),
-    }
-}
-
-fn runtime_config_path(ctx: &ChannelRuntimeContext) -> Option<PathBuf> {
-    ctx.provider_runtime_options
-        .rain_dir
-        .as_ref()
-        .map(|dir| dir.join("config.toml"))
-}
-
-fn runtime_defaults_snapshot(ctx: &ChannelRuntimeContext) -> ChannelRuntimeDefaults {
-    if let Some(config_path) = runtime_config_path(ctx) {
-        let store = runtime_config_store()
-            .lock()
-            .unwrap_or_else(|e| e.into_inner());
-        if let Some(state) = store.get(&config_path) {
-            return state.defaults.clone();
-        }
-    }
-
-    ChannelRuntimeDefaults {
-        default_provider: ctx.default_provider.as_str().to_string(),
-        model: ctx.model.as_str().to_string(),
-        temperature: ctx.temperature,
-        api_key: ctx.api_key.clone(),
-        api_url: ctx.api_url.clone(),
-        reliability: (*ctx.reliability).clone(),
-    }
-}
-
-async fn config_file_stamp(path: &Path) -> Option<ConfigFileStamp> {
-    let metadata = tokio::fs::metadata(path).await.ok()?;
-    let modified = metadata.modified().ok()?;
-    Some(ConfigFileStamp {
-        modified,
-        len: metadata.len(),
-    })
-}
-
-fn decrypt_optional_secret_for_runtime_reload(
-    store: &crate::security::SecretStore,
-    value: &mut Option<String>,
-    field_name: &str,
-) -> Result<()> {
-    if let Some(raw) = value.clone() {
-        if crate::security::SecretStore::is_encrypted(&raw) {
-            *value = Some(
-                store
-                    .decrypt(&raw)
-                    .with_context(|| format!("Failed to decrypt {field_name}"))?,
-            );
-        }
-    }
-    Ok(())
-}
-
-async fn load_runtime_defaults_from_config_file(path: &Path) -> Result<ChannelRuntimeDefaults> {
-    let contents = tokio::fs::read_to_string(path)
-        .await
-        .with_context(|| format!("Failed to read {}", path.display()))?;
-    let mut parsed: Config =
-        toml::from_str(&contents).with_context(|| format!("Failed to parse {}", path.display()))?;
-    parsed.config_path = path.to_path_buf();
-
-    if let Some(rain_dir) = path.parent() {
-        let store = crate::security::SecretStore::new(rain_dir, parsed.secrets.encrypt);
-        decrypt_optional_secret_for_runtime_reload(&store, &mut parsed.api_key, "config.api_key")?;
-        // Decrypt TTS provider API keys for runtime reload
-        if let Some(ref mut openai) = parsed.tts.openai {
-            decrypt_optional_secret_for_runtime_reload(
-                &store,
-                &mut openai.api_key,
-                "config.tts.openai.api_key",
-            )?;
-        }
-        if let Some(ref mut elevenlabs) = parsed.tts.elevenlabs {
-            decrypt_optional_secret_for_runtime_reload(
-                &store,
-                &mut elevenlabs.api_key,
-                "config.tts.elevenlabs.api_key",
-            )?;
-        }
-        if let Some(ref mut google) = parsed.tts.google {
-            decrypt_optional_secret_for_runtime_reload(
-                &store,
-                &mut google.api_key,
-                "config.tts.google.api_key",
-            )?;
-        }
-    }
-
-    parsed.apply_env_overrides();
-    Ok(runtime_defaults_from_config(&parsed))
-}
-
-async fn maybe_apply_runtime_config_update(ctx: &ChannelRuntimeContext) -> Result<()> {
-    let Some(config_path) = runtime_config_path(ctx) else {
-        return Ok(());
-    };
-
-    let Some(stamp) = config_file_stamp(&config_path).await else {
-        return Ok(());
-    };
-
-    {
-        let store = runtime_config_store()
-            .lock()
-            .unwrap_or_else(|e| e.into_inner());
-        if let Some(state) = store.get(&config_path) {
-            if state.last_applied_stamp == Some(stamp) {
-                return Ok(());
-            }
-        }
-    }
-
-    let next_defaults = load_runtime_defaults_from_config_file(&config_path).await?;
-    let next_default_provider = providers::create_resilient_provider_with_options(
-        &next_defaults.default_provider,
-        next_defaults.api_key.as_deref(),
-        next_defaults.api_url.as_deref(),
-        &next_defaults.reliability,
-        &ctx.provider_runtime_options,
-    )?;
-    let next_default_provider: Arc<dyn Provider> = Arc::from(next_default_provider);
-
-    if let Err(err) = next_default_provider.warmup().await {
-        if crate::providers::reliable::is_non_retryable(&err) {
-            tracing::warn!(
-                provider = %next_defaults.default_provider,
-                model = %next_defaults.model,
-                "Rejecting config reload: model not available (non-retryable): {err}"
-            );
-            return Ok(());
-        }
-        tracing::warn!(
-            provider = %next_defaults.default_provider,
-            "Provider warmup failed after config reload (retryable, applying anyway): {err}"
-        );
-    }
-
-    {
-        let mut cache = ctx.provider_cache.lock().unwrap_or_else(|e| e.into_inner());
-        cache.clear();
-        cache.insert(
-            next_defaults.default_provider.clone(),
-            Arc::clone(&next_default_provider),
-        );
-    }
-
-    {
-        let mut store = runtime_config_store()
-            .lock()
-            .unwrap_or_else(|e| e.into_inner());
-        store.insert(
-            config_path.clone(),
-            RuntimeConfigState {
-                defaults: next_defaults.clone(),
-                last_applied_stamp: Some(stamp),
-            },
-        );
-    }
-
-    tracing::info!(
-        path = %config_path.display(),
-        provider = %next_defaults.default_provider,
-        model = %next_defaults.model,
-        temperature = next_defaults.temperature,
-        "Applied updated channel runtime config from disk"
-    );
-
-    Ok(())
-}
-
-fn default_route_selection(ctx: &ChannelRuntimeContext) -> ChannelRouteSelection {
-    let defaults = runtime_defaults_snapshot(ctx);
-    ChannelRouteSelection {
-        provider: defaults.default_provider,
-        model: defaults.model,
-        api_key: None,
-    }
-}
-
-fn get_route_selection(ctx: &ChannelRuntimeContext, sender_key: &str) -> ChannelRouteSelection {
-    ctx.route_overrides
-        .lock()
-        .unwrap_or_else(|e| e.into_inner())
-        .get(sender_key)
-        .cloned()
-        .unwrap_or_else(|| default_route_selection(ctx))
-}
-
-fn set_route_selection(ctx: &ChannelRuntimeContext, sender_key: &str, next: ChannelRouteSelection) {
-    let default_route = default_route_selection(ctx);
-    let mut routes = ctx
-        .route_overrides
-        .lock()
-        .unwrap_or_else(|e| e.into_inner());
-    if next == default_route {
-        routes.remove(sender_key);
-    } else {
-        routes.insert(sender_key.to_string(), next);
-    }
-}
-
-fn clear_sender_history(ctx: &ChannelRuntimeContext, sender_key: &str) {
-    ctx.conversation_histories
-        .lock()
-        .unwrap_or_else(|e| e.into_inner())
-        .remove(sender_key);
-}
-
-fn mark_sender_for_new_session(ctx: &ChannelRuntimeContext, sender_key: &str) {
-    ctx.pending_new_sessions
-        .lock()
-        .unwrap_or_else(|e| e.into_inner())
-        .insert(sender_key.to_string());
-}
-
-fn take_pending_new_session(ctx: &ChannelRuntimeContext, sender_key: &str) -> bool {
-    ctx.pending_new_sessions
-        .lock()
-        .unwrap_or_else(|e| e.into_inner())
-        .remove(sender_key)
 }
 
 fn replace_available_skills_section(base_prompt: &str, refreshed_skills: &str) -> String {
